requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
motor==3.3.1
pymongo==4.5.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
WeasyPrint==60.1
pydyf==0.8.0
Jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
email-validator==2.0.0
qrcode[pil]==7.4.2
lxml==4.9.3
